% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_get_greeks_single.R
\name{bc_get_greeks_single}
\alias{bc_get_greeks_single}
\title{bc_get_greeks_single}
\usage{
bc_get_greeks_single(
  webdriver,
  symbol,
  date_val,
  near_or_all = c("near", "all"),
  download_dir
)
}
\arguments{
\item{webdriver}{\itemize{
\item The webdriver used (use chrome_driver function)
}}

\item{symbol}{\itemize{
\item The stock symbol to be downloaded
}}

\item{date_val}{\itemize{
\item The date to pull, plus whether it is a 'monthly' (m)
or 'weekly' (w) list
}}

\item{near_or_all}{\itemize{
\item An option to specify whether to downloadd all records
or just the 'near the money' records
}}

\item{download_dir}{\itemize{
\item The directory where the file will be downloaded.
}}
}
\value{
A downloaded file from barcharts.com in the specified folder
}
\description{
This function will download a single greeks file from barcharts.com and will
do basic data manipulation (remove footer row, no commas in the 'Open Int' column)
}
