% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_login.R
\name{bc_login}
\alias{bc_login}
\title{bc_login}
\usage{
bc_login(webdriver, username, pw)
}
\arguments{
\item{webdriver}{\itemize{
\item The webdriver used (use chrome_driver function)
}}

\item{username}{\itemize{
\item The login email address
}}

\item{pw}{\itemize{
\item The password associated with the login
}}
}
\value{
The supplied webdriver will now be logged in to barchart.com
}
\description{
After opening a chrome driver, this function navigates to the barcharts.com
webpage and logs a user in. Username and password must be specified.
}
