% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome_driver.R
\name{chrome_driver}
\alias{chrome_driver}
\title{chrome_driver}
\usage{
chrome_driver(download_dir, timeout = "3600")
}
\arguments{
\item{download_dir}{\itemize{
\item Where downloads should go by default
}}

\item{timeout}{\itemize{
\item The amount of time (in seconds) before webpage timeout
}}
}
\value{
A chrome driver instance with browser opened.
}
\description{
This starts a chrome driver instance in Python
}
